Make -Wall -pedantic happy

+++ parser.cpp
@@ -115,7 +115,7 @@ struct Init {
 
     precedences.resize(OperatorClass::Tertiary + 1);
 
-    for (int prec = 0; prec < operatorClasses.size(); prec++) {
+    for (size_t prec = 0; prec < operatorClasses.size(); prec++) {
       for (auto curr : operatorClasses[prec].ops) {
         precedences[operatorClasses[prec].type][curr] = prec;
       }
